+++ OOP_Project/MetricsJSONModels/ListOfUsersAssignedAndTheirCompletenessPercentage.java
@@ -10,7 +10,7 @@ public class ListOfUsersAssignedAndTheirCompletenessPercentage {
     private Integer userId;
     @SerializedName("completenessPercentage")
     @Expose
-    private Integer completenessPercentage;
+    private Double completenessPercentage;
 
     public Integer getUserId() {
         return userId;
@@ -20,11 +20,11 @@ public class ListOfUsersAssignedAndTheirCompletenessPercentage {
         this.userId = userId;
     }
 
-    public Integer getCompletenessPercentage() {
+    public Double getCompletenessPercentage() {
         return completenessPercentage;
     }
 
-    public void setCompletenessPercentage(Integer completenessPercentage) {
+    public void setCompletenessPercentage(Double completenessPercentage) {
         this.completenessPercentage = completenessPercentage;
     }
